feat: Add AOP support for bsrvcore.
AOP is a modern web developing method in mmodern development process.
With the help of "aspect", developers can process with some helper
aspects like log aspects and SSE heart beat aspect and focus the
developer on core logics.

AOP support is mentioned in the initial plan but due to some small
mistakes, the UML design has not included this function in. Thus, this
function is added to the project and the UML diagram is well modified.